Add files via upload
"This marks the beginning of my exciting journey with Codsoft! 🚀 My second task as an intern, thoughtfully provided by the Codsoft team, is to 'Calculate total marks, average percentage, and grade' in Java. 📊 I'm thrilled to learn, grow, and contribute to amazing projects during this internship. Let's make the most of this opportunity! 💪 #InternshipKickoff #CodsoftIntern #SecondTask #java"